requirements.txt
ripe.atlas.cousteau
pyspark
pyspark-stubs
requests

